done makefile and add struct

+++ libft/src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: preed <preed@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/09 17:43:40 by preed             #+#    #+#             */
-/*   Updated: 2022/03/22 20:26:00 by preed            ###   ########.fr       */
+/*   Updated: 2022/07/18 16:09:02 by lgarrosh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 //counts symbols in line
 //does not count the last null-terminating character
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	size_t	i;
